move TestTaskModule to tests.common.taskmodule
